Build image from python:2-alpine and removed ONBUILD statements
Using an alpine linux base for this image reduces the size of the image,
although we do have to add build utils back in for the python
dependencies. As we reduce those dependencies, this image will get
smaller.

Removing the ONBUILD statements from the Dockerfile simplifies using
this image as a base to build other images from. Documentation will need
to be added for what needs to be mounted where