Rewrite Dockerfile and docker_launch to enable container-only deploys
Right now, users have to install terraform and ansible, and their
dependencies, in order to user Mantl. Vagrant is a demo/test-drive
option, but not for other use cases. After the installation(s), then
they have to run multiple commands in order to get a cluster up.

What I'm working towards is a way for users to be able to get started
with Mantl in one or two steps, that removes the need to install
dependenies. This commit reworks Dockerfile and docker_launch.sh, and in
the tests that I have done, it can accomplish this.

The ONBUILD statements are replaced with logic in
docker_launch.sh to check for user-provided files in a particular place.
If these files don't exist, then we copy/generate the files. (An
exception to this is SSH keys.) When users mount a host volume in the
MANTL_CONFIG_DIR, they will get copies of all of the generated files
in that volume, which they are free to customize further.

TODO: documentation, test without mounting preexisting mantl directory