Add SSH_KEY env to Dockerfile, refactored directives
Some of the logic in docker_launch.sh script relied on SSH_KEY in the
env, and I removed it for some reason. Originally, it was set in the
launch script, so now I'm setting it in the Dockerfile.

Also, I refactored the Dockerfile. Some of the ENV lines are separated
out for readability, and I collapsed a couple of RUN lines into one RUN
line, to reduce space.