Remove SSH_KEY env from Dockerfile since it's semantics will change
I'm changing the semantics of the SSH_KEY env. Since all the SSH tools
assume that the key is at ~/.ssh/id_rsa, I decided that it would be
simpler to symlink any user provided keys to that location. Therefore,
the SSH_KEY env var will refer to the location of the user-provided key