Migrate testing to use Docker-based deployments, with many changes
1. the Travis file has lost many of its directives in favor with moving
them over to a new home in a setup script
2. The docker_launch script has also lost its setup logic, and now
contains only the steps necessary to get a cluster up and running
3. the docker_setup python script takes care of getting the container
into proper state. As long as you mount a volume onto /local, your
config files will be symlinked to their proper locations. Otherwise,
they are generated for you, and saved to the mounted volume.