Move travis build steps into a "switchboard" run-tests script
Our CI setup was complex to start with, and the complexity is
increasing. To help with this, I moved much of the setup logic from the
travis config file to a separate script. This will allow us to skip
certain builds (such as documentation) and it will handle more complex
cases, such as accessing a jump box, only trigger cloud deploys in
certain situations, and so on.

This is mostly a refactoring, with little logic changes.