Add relative path to tf files found by docker_setup.py
Inside the list comprehension that generates the list of terraform files
to symlink for docker setup, I misunderstood what was returned by
os.listdir. I believed that the relative path was collected into the
list. However, this was not the case, and therefore the symlinks that
were generated from that function were self-referential and broken.
Now, the os.relpath function is called for each member of the list
during the comprehension, so the behavior is now as I expected.