Add keygen/link function to launch & build-cluster
We need ssh keys to be symlinked for launch and testing, but we don't
want to run the whole setup script each time. So, I added a call to the
ssh key gen/link function to build-cluster.py and docker_launch.sh.
Assuming that the keys have already been generated from a run of
docker_setup.py, all that function does is symlink them.