Dereference secret envs with ${} syntax
After I was done debugging the docker setup script, we now have an error
where the secret env vars are not being injected into the docker
containers properly. I took a second look at the travis matrix, and I
realized that the secrets where referenced in $var syntax, and I think
that, similar to the pwd issue, they are not being expanded as I expect.
Therefore, I'm wrapping them in ${} to try to get them deferenced
properly.