Expand secrets with ${}, rm debugging, combine -e's
The secret env vars aren't expanded properly, so I need to use the
correct bash syntax to do so. I'm wrapping the vars in ${}, which I hope
will behave like TF_VAR_build_number. The next thing is to use an eval
command if this doesn't work.

Also, I removed a line that I inserted for debugging purposes.