Using ssh-agent with t option to run test cmds
ssh-agent has a 'lifetime option' where you can pass in a command and
args. The agent then only runs for a specified time, and the command is
forked. My other efforts to get the agent active for the test runs have
failed, so this is my next try. If this doesn't work, I will investigate
other options.