Readd sh invocation to entrypoint, fix path for setup
With Brian's advice, we add the sh invocation back into the entrypoint
of the Dockerfile, so that we can use the container as an executable.
This allows us to use the benefits of the ssh-agent lifetime feature,
and keep the container flexible.

The other change in here is that I fixed the relpath for the
docker_setup script in destroy_build.sh. With this change, the builds
will be destroyed properly once again.