Install curl to image, rm renaming of security.yml
To optimize for the size of the docker image, I have apk delete the
package. curl is a dependency of git, so when I removed that package, I
also removed curl. Therefore, we install it separately.

Secondly, I had logic in the security_setup wrapper function to move the
generated security file into MANTL_CONFIG_DIR, then symlink it. However,
Travis threw an error. Since that functionality is for users, and not
CI, I elected to remove that functionality, and add it in later if
needed.