Import check_output, print out repo_slug
The check_output function wasn't properly imported into docker.py, so I
fixed that in the import section. Additionally, I added a line looking
at the TRAVIS_REPO_SLUG env var: according to their docs, it should have
the slug (repo_owner/repo_name) of the build. If that is the case, then
we can inspect that string, and have different behaviors for forks based
on that.