Add shlex split and git, port grep exclusion
I added the shlex split function to the script, so the git command can
get interpreted more accurately. I had to add the git package back into
the image (I orginally removed it to save on space).

The commit_range_cmd required a subshell, and using a subshell for a
subprocess is risky. I ported the grep invocation to python if/elif/else
block, with logging.info for files detected. The log lines are probably
too verbose.