Setup func -> ci_setup, split link tf files
I'm quickly iterating through these changes for travis, and I'm finding
that some of the assumptions that I made for a user are not working out
well for travis.

Because we are focusing on getting CI working before we get this working
for end users, I decided that we are going to have commands specifically
for CI. Those commands are prefixed with "ci-". I want to explore the
option of detecting a "CI" env var, but I felt that this was a
straighter path to our goal.

The TF files need to be symlinked during the build and destroy steps,
not just during setup.  @siddharthist made a good point about the
terraform files without their extension. The "clever" way that I
circumvented terraform grabbing multiple terraform files was clumsy, so
I replaced it with a better approach. I added logic to the setup
function in docker.py to symlink the file defined in TERRAFORM_FILE if
the env exists.

Additionally, I added OS_PRT to the secret envs. This represents
CiscoCloud's secret ssh port number.