Clone mantl for each build
For testing against OpenStack, we should have separate clones of mantl
for each test run, to avoid collisions during concurrent runs. The
remote command for this now clones the repository into a new directory
that's the same name as the commit, checks out that commit, runs tests,
then removes the directory after tests are completed.